rag/pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.20",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pyjwt>=2.8.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.32.0",
]

[project.scripts]
start = "src.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
